Support smtaddr2line tool to parse memoryleak file into source line
if user compile with -DUSE_WTF_SYMBOLIZE=1,
the bactrace will be demanged with function name.
if not, user can use smtaddr2line to view more info of memoryleak

it would better compile with -g option when using smtaddr2line tool,
so that the memory leak can be directed the source line.